parts strewn all over the floor
